# PERF: Cython version of Python _TIMEPAT regexp in parsing.pyx