- [x] closes #N/A
- [ ] tests added / passed
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry

`asv continuous -f 1.05 origin/master HEAD -b ^io.csv.ReadCSVParseSpecialDate -a warmup_time=2 -a sample_time=2`:

master | patch | ratio | test_name
-|-|-|-
276±1ms|          262±2ms|     0.95|  io.csv.ReadCSVParseSpecialDate.time_read_special_date('hm')
16.6±0.2ms|      14.2±0.08ms|     0.85|  io.csv.ReadCSVParseSpecialDate.time_read_special_date('mdY')
42.7±0.9ms|       35.0±0.2ms |    0.82|  io.csv.ReadCSVParseSpecialDate.time_read_special_date('mY')