@jreback 2019-04-24
also merge master; ci has been fixed