@vnlitvin 2019-04-24
> also merge master; ci has been fixed

I've rebased instead, I personally find rebased history cleaner than merged.