@jreback 2019-04-24
> I've rebased instead, I personally find rebased history cleaner than merged.

I agree :-D

we use merge master as its somwhat friendlier to new users, in any event everything is squashed on merge anyways (so the commit history is only relevant to you)